requirements.txt
# Common
openai
python-dotenv
pytest

# Task 1: AI Architect API
fastapi[all]
uvicorn[standard]
requests
black
sentence-transformers
faiss-cpu

# Task 2: AI Readiness Assessment
streamlit
streamlit-survey
pandas
numpy
matplotlib
plotly
markdown